fix(Review Comments & Session):
Calculate modified lines in the constructor instead of calling it for every discussion.
Remove the deserializer from the `location` field and use ``@JsonProperty` instead.
Don't allow the user to add a new line in the title of an issue or patch.
After the recent update (1bb400c2)  `rad web --json` command doesnt exists. Instead use `rad web --no-open`  in order to authenticate the user and get the sessions for the apis

Signed-off-by: Stelios Mavrommatakis <steliosmavr@cytech.gr>