fix: Authenticate User
This commit implements the `AuthService.class` to manage Radicle authentication and signing sessions with the user's private key. The class includes logic to check if the user's password is stored in IntelliJ; if so, it signs the keys without requiring the user to input their password. If the password is not stored, then show a dialog to the user to input their password and stored for future use.

Signed-off-by: Stelios Mavrommatakis <steliosmavr@cytech.gr>