fix(Null exception): View file diff in a patch
When the user tried to view a file diff the was an error at  the calculation of the modified lines and we were getting null exception in some cases. Fix this bug by adjusting the modifiedLines method and also enable the user to add review comments in both files when viewing the diff.

Signed-off-by: stelios maurommatakis <steliosmavr@cytech.gr>