fix(patches): Fix diff for unknown commit
When trying to calculate commits for a patch revision,
the head may not exist locally (even after fetch).

In that case, calculating commits with `git log xxx...yyy`
would produce an error about `Invalid symmetric difference expression`.

Since we have already fetched, there's nothing better to do,
so catch the error and keep on calculating commits for other revisions.

This way, the patch can still be displayed,
even if a specific revision was _inaccessible_.

Signed-off-by: jchrist <jchrist@cytech.gr>