fix(inline comments): Show  comments from previous revisions
Always display comments that belong to the latest revision or are on a line that did not change in the next revision. If a comment does not belong to the current revision, add the word "OUTDATED" next to the author of the comment to indicate that it is from a previous revision. In the timeline for an inline comment, show the file name and the line number where the comment was made.

Signed-off-by: stelios maurommatakis <steliosmavr@cytech.gr>